+++ packages/cubing-core/mod.ts
@@ -28,7 +28,7 @@
  * @module
  */
 
-export const VERSION = "0.3.2";
+export const VERSION = "0.3.3";
 
 export {
   formatAlg,
@@ -36,6 +36,8 @@ export {
   invert,
   isDouble,
   isPrime,
+  mergeAdjacent,
+  mergeAdjacentWithPrefixes,
   type Move,
   type MoveFamily,
   NotationError,